Proposal Detail: Refactor general component structure and do UI tweaks
